refactor: reorganize FileBrowser around the dialog's actual structure
